feat: add dtos, mappers

+++ Spring2/src/main/groovy/com/example/productapp/controller/CategoryController.java
@@ -1,6 +1,8 @@
 package com.example.productapp.controller;
 
+import com.example.productapp.dto.CategoryDTO;
 import com.example.productapp.entity.Category;
+import com.example.productapp.mapper.CategoryMapper;
 import com.example.productapp.service.CategoryService;
 import com.example.productapp.validators.CategoryValidator;
 import org.slf4j.Logger;
@@ -35,23 +37,31 @@ public class CategoryController {
     }
 
     @GetMapping
-    public List<Category> getAllCategories() {
+    public List<CategoryDTO> getAllCategories() {
         return categoryService.getAllCategories();
     }
 
+
     @PostMapping
-    public ResponseEntity<?> createCategory(@RequestBody @Validated Category category, BindingResult result) {
-        log.info("Creating category: {}", category);
+    public ResponseEntity<?> createCategory(@RequestBody @Validated CategoryDTO categoryDTO, BindingResult result) {
+        log.info("Creating category: {}", categoryDTO);
         log.info("Binding result: {}", result);
+
         if (result.hasErrors()) {
             return ResponseEntity.badRequest().body(result.getAllErrors());
         }
+
+        Category category = CategoryMapper.toEntity(categoryDTO);
         Category savedCategory = categoryService.saveCategory(category);
-        return ResponseEntity.status(HttpStatus.CREATED).body(savedCategory);
+        return ResponseEntity.status(HttpStatus.CREATED).body(CategoryMapper.toDTO(savedCategory));
     }
+
 
     @PutMapping("/{id}")
     public ResponseEntity<?> updateCategory(@PathVariable("id") Long id, @RequestBody @Validated Category category, BindingResult result) {
+        if(categoryService.getAllCategories().stream().noneMatch(c -> c.getId().equals(id))){
+            return ResponseEntity.badRequest().body("Category with id " + id + " not found");
+        }
         if (result.hasErrors()) {
             return ResponseEntity.badRequest().body(result.getAllErrors());
         }
@@ -61,7 +71,11 @@ public class CategoryController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCategory(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteCategory(@PathVariable("id") Long id) {
+        if(categoryService.getAllCategories().stream().noneMatch(c -> c.getId().equals(id))){
+            return ResponseEntity.badRequest().body("Category with id " + id + " not found");
+        }
         categoryService.deleteCategory(id);
+        return ResponseEntity.ok().body("Category with id " + id + " deleted");
     }
 }
